refactor: Split up app file into properly structured files
