+++ application/__init__.py
@@ -1,1 +1,2 @@
-import application.room
+from application.flask import app, socket_app, rooms
+import application.endpoints
