+++ application/endpoints/__init__.py
@@ -1,0 +1,3 @@
+import application.endpoints.home
+import application.endpoints.room
+import application.endpoints.socketio
