+++ application/model/__init__.py
@@ -1,0 +1,1 @@
+import application.model.room
